backend/pyproject.toml
[tool.poetry]
name = "unloading"
version = "0.1.0"
description = ""
authors = ["Constantine Putilov <putilov.constantine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.27.2"
psycopg2 = "^2.9.9"
gunicorn = "^22.0.0"
celery = "^5.4.0"
redis = "^5.0.6"
openpyxl = "^3.1.4"
pandas = "^2.2.2"


[tool.poetry.group.test.dependencies]
factory-boy = "^3.3.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
django-debug-toolbar = "4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
